app/requirements.txt
django==3.2
flake8==4.0
django-debug-toolbar==3.4.0
python-dotenv==0.21.0
django-split-settings==1.1.0
psycopg2==2.9.3
pytest==7.1.3
uwsgi==2.0.20
django-cors-headers==3.13.0
